feat: implement component organization again for blog posts page
